+++ src/Datasets/Unlabeled.php
@@ -652,7 +652,7 @@ class Unlabeled extends Dataset
         }
 
         array_unshift($table, $header);
-        $columnSize = floor($tCols) / count($table[0]);
+        $columnSize = (int) floor($tCols) / count($table[0]);
 
         return Console::table($table, $columnSize);
     }
